tools/requirements.txt
pwgen
python-decouple
requests
unidecode
